Split new cluster into separate function
Trying to make controller.go less daunting.
